Created Puzzle ball functionality and added ladder
Created a puzzle ball with 3 controllers (lever, joystick and wheel). The lever needs to be on so the joystick and wheel can actually move. The joystick moves the ball forward/backward and left/right. The wheel moves it up/down. Code for the functionality is in the PuzzleBallController class. Also added a ladder with materials in the main hall