Created  puzzle maze with ball
Created maze with ball. The ball is controlled by a jotstick and a wheel knob . The joustick moves the ball left/right and forward/backwards. The wheel moves the ball up/down. A second joystick moves the actual maze in a rotation. Player must get the ball to the end of the maze. Colour material was also added to the maze